Math: ability to construct all types without initializing the contents.
Useful for squeezing out last bits of performance, e.g. in this case:

    Vector3 a;
    a[0] = something++;
    a[1] = something++;
    a[2] = something++;

In the code all elements are first zeroed out and then overwritten
later, thus it might be good to avoid the zero-initialization:

    Vector3 a{Math::NoInit};
    a[0] = something++;
    a[1] = something++;
    a[2] = something++;

This will of course be more useful in far larger data types and arrays
of these.