Compressed image support, part 6: Compressed* image classes.
Added (now empty) AbstractCompressedImage class that inherits (now also
empty) AbstractImage class.

Added CompressedBufferImage, CompressedImage and CompressedImageView
classes, which are just copies of BufferImage, Image and ImageView
classes with format/type pair replaced by just format, but they
additionally need data size parameter.

Because of different use cases in Trade, the Trade::ImageData class now
handles both uncompressed and compressed format, checking the API
usage with runtime assertions. The reason for this is that a material
could just reference a image file by ID and we need to be able to
extract image of that ID without prior knowledge whether it is
compressed or not. Requiring prior knowledge of image format from the
user would make both the API and the usage far more complicated than
having Trade::ImageData which handles both cases.

On the other hand, the Image*/CompressedImage* distinction is done for
easier usage and type-safe APIs in all other cases.