Store a copy instead of pointer to Buffer in Mesh.
Currently the user had to ensure that buffers added to mesh were not
moved at all, which was very annoying, basically each one of them had to
be allocated on heap. Now the Mesh stores a weak copy (yes, really) using
Buffer::wrap() with no deletion on destruction, so the original instance
can be freely moved around without any fear of crash.

Thanks to @Squareys for the original idea/request about wrap() functions,
really useful part of the API.