Trade: improve AbstractImageConverter API.
Provide a way to convert compressed images to data/file (i.e. saving
DXT5-compressed image data to DDS file), improve feature flags so that
the plugin can properly advertise what's supported (for example some
plugin may just be able to compress RGBA to DXT5, but not to save that to
DDS file).

This is backwards-incompatible API breakage (renamed enum value), but
because the original API wasn't in any official release yet, I'm not
doing any deprecation and backwards compatibility.