SceneGraph: document projectionSize() usage.
I got like five e-mails about this already, putting that in the docs so I
don't have to invent that every time again.

This is very specific to many transformation/projection properties
(origin, Y up, 2D/3D and whatnot) and thus I'm not adding any convenience
function to calculate that.