Platform: do the core context workaround also for AMD.
AMD is behaving the same as NVidia (at least on Windows) -- when
creating core context with minimum specified version set to 3.1, it
forces that version instead of going with the largest available version,
which, again, is pretty useless behavior.

Enabling that on both Linux and Windows, the behavior is confirmed on
Windows but I bet it's doing the same on Linux.