AbstractFramebuffer: Use correct target for GLES
OpenGL ES without GL_APPLE_framebuffer_multisample doesn't
support multiple draw or read bindings, so we need to use
GL_FRAMEBUFFER for that.

It seems as though the Magnum convention is to use use
FramebufferTarget::Draw for read-draw bindings, so Read and Draw
have been aliased to GL_FRAMEBUFFER in this case.