modules: don't unset MAGNUM_*APPLICATION_* vars if they stay the same.
Use case: user is calling find_package(Magnum REQUIRED Sdl2Application) in
a toplevel dir and in a subdir. The second call would unset the
MAGNUM_APPLICATION_LIBRARIES which is not correct, because it still is just
that one application.

Also, now instead of unsetting the variables are set to -NOTFOUND so they
produce an error when used as libraries/include directories for a target.