Math: better test cases for {*Complex,Quaternion}::isNormalized().
The expectation is that the values are considered normalized only if the
difference is small enough. This should have been tested since the
beginning, but instead this was waved away with a dumb test case testing
obviously denormalized value and obviously normalized value.

The test fails for DualQuaternion with large translation values (as
expected). Will be fixed in following commits.