CMake: better implicit behavior for WITH_WINDOWLESS{EGL,GLX}APPLICATION.
I found a way how to hack cmake_dependent_option() beyond just
specifying a list of things AND-ed togeter -- if one uses spaces instead
of `;`, all the contents are passed to a single if() statement and it's
possible to use parentheses, AND, OR and all other stuff.

Now, WindowlessEglApplication is implicitly enabled if targeting
non-desktop ES and magnum-info utility is enabled, because magnum-info
depends on it in that case. In all other cases it is implicitly disabled
with ability to enable it.

WindowlessGlxApplication is now available only when targeting desktop GL
or desktop GLES and is implicitly enabled when magnum-info or any other
GL utility is enabled.