Math: functions for half-float (un)packing.
Most of the code is in the actual test where I'm comparing and
benchmarking three different implementations (a
naive/straightforward/ground-truth one, the chosen one and a fast though
cache-spilling table-based one) to ensure the behavior is consistent
across all of them and that the performance is within reasonable bounds.

The Corrade::TestSuite benchmarking stuff needs serious improvements,
though.