Platform: proper NoCreate constructors for windowed classes.
The move away from `nullptr` to NoCreate for constructing an application
without creating OpenGL context was done quite some time ago for
windowless application, but for some weird reason it was never done for
windowed apps. Now made this consistent.

The old `nullptr`-based constructor is still present, but marked as
deprecated and due to be removed in some future release.