Buffer: deprecate templated overloads of data(), subData() and map().
They were utterly confusing, as it was completely unclear what the units
of offset/size parameters are, whether byte sizes or element counts (and
moreover, some of these APIs had offset in bytes and size in count and
some not). All of those are deprecated now, with hinting the user to
convert to non-templated APIs in combination with
Containers::arrayCast(). Moreover, the non-templated range map()
function doesn't return just void* anymore, but a properly sized
ArrayView<char>. The old map() (which doesn't take range) still returns
just a pointer (but also a char* instead of void* for consistency), as
getting size there is non-trivial (and impossible on old ES/WebGL).
The switch to ArrayView might be a source breaking change, but I
silently hope that everyone was just using the templated functions
anyway (that are deprecated now). So, in short, this was before:

   T* a = buf.map<T>(0, size_in_what_i_have_no_idea);

And this is now, with proper size safety and clear API:

   ArrayView<T> a = Containers::arrayCast<T>(buf.map(0, size_in_bytes);

The deprecated APIs will be removed at some point in the future, as
usual.