Platform: un-deprecate creation of Platform::Context without passing args.
I deprecated that because I wanted to be sure that I properly update all
my code to actually propagate the options. But now I see that this is
actually a valid case -- the engine being in a thread or in some other
way hidden from the outside *shouldn't* need access to argc/argv.

Also updated the docs to say it's completely harmless to propagate the
args.