Converting APIs returning std::optional to Containers::Optional.
And removing the bundled std::optional implementation. This finally
makes this library compatible with C++17. Since this would be a huge
backwards-incompatible change that would make everyone angry, the
following had to be done in case both CORRADE_BUILD_DEPRECATED and
MAGNUM_BUILD_DEPRECATED is defined:

 * Under C++11 and C++14, Containers::Optional / Containers::NullOpt is
   aliased to std::optional / std::nullopt. This is no worse than the
   state before, when we also provided these symbols.
 * Under C++17, where standard <optional> header is available,
   Containers::Optional provides implicit conversion to it. Only one-way
   conversion is supported, as there was fortunately no Magnum API that
   took std::optional via parameter, and there might be some corner
   cases that this doesn't cover. The goal is to have all examples
   compiling with the old API, at least.
 * There's a new test especially for this, which checks that both the
   C++11 and C++17 ways of doing things work as they should.

The typedef and conversion is marked as deprecated, so it will spit out
many warnings to push users to upgrade. I hope I can completely remove
this mess soon :/