Shaders: rename texture binding functions from set*() to bind*().
This better reflects that the functions modify a global state instead of
a shader-local state and so rebinding may be necessary (unlike with
uniforms, which get preserved).

The old set*() functions are now inline aliases to the bind*()
functions, are marked as deprecated and will be removed in some future
release.