Split the OpenGL layer out, pt 10: test the new image APIs.
With the previous commits the original tests passed (which is
desired), but these were using deprecated functionality and not
covering the new stuff. These tests are not using the deprecated
functionality, which means I don't need to build them as part of
the GL library anymore.

The GL::BufferImage test is still using the deprecated
functionality though, in order to check I didn't break anything
by accident.