GL: fix undefined behavior and potential leaks in Mesh.
The change done in 680144f1c511997ee0530ea88bbb95aa20e058c8 was not
properly handling these cases:

 * Mesh(NoCreateT) and wrap() were not constructing the internal vector,
   which blew up when move-assigning another instance.
 * ~Mesh() was not destructing the internal vector if the VAO ID was
   zero or non-owning wrap() was used.

Strangely enough none of these were causing *any* problems for me on
Linux (even ASan was totally quiet and due to the unfortunate
combination of bugs even when I assigned totally random data to the
storage vector). This however blew up on MSVC, assuming there the
implementation is more checked.

Because it's possible to construct Mesh with no GL context available,
the move construction and destruction needs to avoid accessing Context
unless really necessary (it would be also unclear which type of vector
should be constructed if we have no context).

Extended the tests to handle hopefully all the cases.