package/ci: build minimal Vulkan support on Linux and Windows.
Disabling everything that doesn't depend on Vulkan as it's not needed to
be built again -- that's basically everything right now. This needs to
be a separate build for the foreseeable future, because it requires a
newer CMake version for the Find module and it'll probably also require
newer C++ standard.

Thanks to @LB-- for providing me with an oneliner to compile a file with
cl.exe.