Math: fix sclerp() behavior with dual quaternions having the same angle.
The fix done with https://github.com/mosra/magnum/pull/122
(0e05c7289e21b21596161e8f3c30cc3da6f624ec) was not tested properly (see
previous commit) and thus this code path never worked. This properly
lerps the translation part and recombines it with the rotation instead
of interpolating just a part of it.

Also I'm no longer having any "dotResult" that's done only on the
vector part of the rotation, but instead using the full
rotation quaternion dot product. I have no idea why it was done this
way. This branch was also never properly tested -- it'll be with the
introduction of "shortest path" variants in the next commit.