GL: actually fix the VAO / element array buffer binding state interaction.
Fixing the new (and now failing test) from the previous commit. In
setIndexBuffer(), I was resetting the state tracker to a VAO state that
was about to be set in the very next step, and then, when doing that
next step the state tracker "optimized away" the state change because it
thought it was already done (even though it wasn't). Reordering the two
operations fixes it.

In comparison to how this was meant to be done in the original
169031fb7bf30e2c98c81cbf865aa5a4662aa716, the new way should do the same
but additionally avoid a bunch of redundant state calls. Let's hope no
more bugs related to this appear.