DebugTools: don't try to use system plugin dirs in CompareImage test.
Due to that, all the tests got skipped on the CI and moreover when ABI
was broken, tests were failing hard due to ABI mismatches with system
installed plugins. No good.

Some variants still need to access the system plugin dir (such as checks
for compressed data), but those are minimal. Sanitized them nevertheless
so a broken system-installed plugin doesn't break the test.